backend_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
loguru>=0.7.0
python-multipart>=0.0.6
httpx>=0.24.0
pytest>=7.0.0
pytest-cov>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.1.0
email-validator>=2.0.0
